begin moving parser/scanner code into separate dirs
